refactor: Update .env file to add development flag for backend file usage
Update the .env file to include a new flag, SD_BACKEND_FILE, which is set to false by default. This flag is used for development purposes and indicates whether the backend file should be used instead of making an API call to retrieve status data.

refactor: Update Status.tsx to conditionally fetch status data from API or mock file

In the Status.tsx file, modify the useRequest hook to conditionally fetch the status data. If the SD_BACKEND_FILE flag in the .env file is set to true, the data is imported from the mock.json file. Otherwise, the data is fetched from the API endpoint.

add: Create mock.json file for testing purposes

Create a new mock.json file in the Services directory. This file is used for testing purposes and currently does not contain any data.