feat: Adding a Static Web Apps Workflow
A workflow called Static Web Apps has been added to support the building and deployment of static web apps. This workflow is triggered when pushing to the master branch.

The workflow consists of a job called build_deploy that runs on the latest version of Ubuntu. The goal of the job is to build and deploy the application.

Specific steps include:
- Checking out code using the actions/checkout@v4 action
- Setting up the Node.js environment using the actions/setup-node@v4 action
- Install the latest version of pnpm using the action pnpm/action-setup@v4.
- Modify the .env file to change the value of SD_BACKEND_FILE from false to true.
- Download the mock.json file using the curl command
- Use the pnpm command to install the dependencies and build the application.
- Use the peaceiris/actions-gh-pages@v4 action to publish the built files to the gh-pages branch, and set the custom domain name to sd3.eco.tsi-dev.otc-service.com.

The purpose of this workflow is to support automated build and deployment of static web applications.